api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "django-ninja>=1.4.3",
    "gql[requests]>=3.5.3",
    "gunicorn>=23.0.0",
    "openai>=1.97.1",
    "psycopg[binary,pool]>=3.2.9",
    "python-dateutil>=2.9.0.post0",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
